feat: add liquibase

+++ src/main/java/com/javaspring/proj/Repository/BidRepository.java
@@ -3,7 +3,7 @@ package com.javaspring.proj.Repository;
 import com.javaspring.proj.model.Bid;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BidRepository extends CrudRepository<Bid, Integer> {
+public interface BidRepository extends CrudRepository<Bid, Long> {
     Iterable<Bid> findByTypeAndStatus(String type, String status);
     Iterable<Bid> findByType(String type);
     Iterable<Bid> findByStatusOrderByDateDesc(String status);
